src/clustering/clusterer.py
"""
Article Clustering Module
==========================
Pipeline Stage: Semantic Clustering (Section 5.3)

Groups articles covering the same real-world event using unsupervised clustering
on sentence embeddings:
- Primary method: HDBSCAN (min_cluster_size=3, metric='cosine')
- Baseline comparison: K-Means
- Evaluation: Silhouette Score + manual inspection
- Singleton clusters and noise points are discarded

Also handles multi-document input construction (Section 5.4):
- Concatenates articles within each cluster ordered by publication time
- Truncates to BART's 1024-token input limit
- Prioritizes first 2-3 sentences per article for maximum coverage
"""

import json
import os
from typing import Any

import hdbscan
import numpy as np
import yaml
from sklearn.cluster import KMeans
from sklearn.metrics import silhouette_score


def load_config(config_path: str = "configs/config.yaml") -> dict:
    """Load project configuration from YAML file."""
    with open(config_path, "r") as f:
        return yaml.safe_load(f)


def cluster_hdbscan(
    embeddings: np.ndarray,
    min_cluster_size: int = 3,
    metric: str = "cosine",
) -> np.ndarray:
    """
    Cluster embeddings using HDBSCAN.

    Args:
        embeddings: NumPy array of shape (n_articles, embedding_dim).
        min_cluster_size: Minimum number of articles to form a cluster.
        metric: Distance metric for HDBSCAN.

    Returns:
        Array of cluster labels (-1 indicates noise/unclustered).
    """
    # TODO: Fit HDBSCAN on embeddings
    # TODO: Return cluster labels
    pass


def cluster_kmeans(
    embeddings: np.ndarray,
    n_clusters: int = 20,
) -> np.ndarray:
    """
    Cluster embeddings using K-Means (baseline comparison).

    Args:
        embeddings: NumPy array of shape (n_articles, embedding_dim).
        n_clusters: Number of clusters for K-Means.

    Returns:
        Array of cluster labels.
    """
    # TODO: Fit K-Means on embeddings
    # TODO: Return cluster labels
    pass


def evaluate_clusters(
    embeddings: np.ndarray, labels: np.ndarray
) -> float:
    """
    Evaluate clustering quality using Silhouette Score.

    Args:
        embeddings: Article embedding vectors.
        labels: Cluster assignments.

    Returns:
        Silhouette score (higher is better, range [-1, 1]).
    """
    # TODO: Compute and return silhouette score
    # TODO: Handle case where all points are noise (labels == -1)
    pass


def group_articles_by_cluster(
    articles: list[dict[str, Any]], labels: np.ndarray
) -> dict[int, list[dict[str, Any]]]:
    """
    Group articles into clusters based on labels.

    Args:
        articles: List of article dictionaries.
        labels: Cluster assignment for each article.

    Returns:
        Dictionary mapping cluster_id -> list of articles in that cluster.
        Noise points (label == -1) and singleton clusters are excluded.
    """
    # TODO: Group articles by their cluster label
    # TODO: Discard noise points (label == -1) and singletons
    pass


def construct_multi_doc_input(
    cluster_articles: list[dict[str, Any]],
    max_tokens: int = 1024,
    sentences_per_article: int = 3,
) -> str:
    """
    Construct a single input text from a cluster of articles (Section 5.4).

    Articles are ordered by publication time. The first N sentences from each
    article are concatenated and truncated to the model's max token limit.

    Args:
        cluster_articles: List of articles in a single cluster.
        max_tokens: Maximum token count for BART input.
        sentences_per_article: Number of leading sentences per article.

    Returns:
        Concatenated, truncated input string for summarization.
    """
    # TODO: Sort articles by publication time
    # TODO: Extract first N sentences from each article
    # TODO: Concatenate and truncate to max_tokens
    pass


def save_clusters(
    clusters: dict[int, list[dict[str, Any]]], output_dir: str
) -> str:
    """
    Save cluster assignments and grouped articles to disk.

    Args:
        clusters: Dictionary mapping cluster_id -> list of articles.
        output_dir: Path to the clusters data directory.

    Returns:
        Path to the saved JSON file.
    """
    # TODO: Save clusters as JSON
    pass


def main():
    """Run the full clustering pipeline."""
    config = load_config()

    # TODO: Load processed articles and embeddings
    # TODO: Run HDBSCAN clustering
    # TODO: Run K-Means clustering (baseline)
    # TODO: Evaluate both methods with silhouette score
    # TODO: Group articles by cluster
    # TODO: Construct multi-document inputs for each cluster
    # TODO: Save clusters to data/clusters/
    # TODO: Print summary (number of clusters, articles per cluster, silhouette scores)
    pass


if __name__ == "__main__":
    main()
